Create LICENSE
I was going to use the CC0 License until I realized it's better to avoid it (I mean jeez, just look at their disclaimer at the top of it).  Apache 2.0 would have been great except umbrella projects can't be created/combined as an Apache 2.0 license if it uses code with a GPL v3 license.  So, better to use GPL v3.